requirements.txt
streamlit==1.32.2
openai
pillow
pandas
python-docx
openpyxl
PyMuPDF
python-pptx
PyPDF2
googletrans
python-dotenv
langchain-community
faiss-cpu
tiktoken
langchain
langchain-openai 
streamlit-chat